Changed get_input().  variable is now unsigned int. It's a multiple of the grid resolution instead of a distance (float).

+++ include/GrANA/utils.hpp
@@ -13,7 +13,7 @@ namespace GrANA {
 
 // Read input.
 auto get_input(int argc, char **argv)
-    -> std::tuple<std::string, float, float, std::string>;
+    -> std::tuple<std::string, float, unsigned int, std::string>;
 
 // Helper function for getting the `int` indices that sort a vector.
 template <typename T>
